# swaggers/contents/swagger.yml
openapi: "3.0.2"
info:
    title: Contents API
    version: "0.1"
    description: 영상/교재/위키피디아/아티클 API
servers:
  - url: http://192.168.153.173:8080/
    description: example server address
...
paths:       
...
  /users/{user_id}/wikis/{wiki_id}:
    get:
      tags:
        - Wikipedia
      summary : 유저의 백과사전 조회
      description : 유저의 북마크 정보를 포함한 백과사전 조회
      parameters:
      - name: user_id
        in: path
        description: 사용자 아이디
        required: true
        schema:
          type: string
          format: string
      - name: wiki_id
        in: path
        description: 백과사전 아이디
        required: true
        schema:
          type: integer
          format: integer
      responses:
        200:
          description: OK
          content:
            json:
              schema:
                type: object
                properties:
                  wikiId:
                    type: integer
                  title: 
                    type: string
                  createDate: 
                    type: string
                  creatorId: 
                    type: string
                  imgSrc: 
                    type: string
                  description: 
                    type: string
                  summary: 
                    type: string
                  source: 
                    type: string
                  bookmark: 
                    type: boolean
                  uks:
                    type: array
                    items:
                      type: string
  /users/{user_id}/wikis:
    get:
      tags:
        - Wikipedia
      summary : 유저의 위키피디아 리스트 조회
      description : 유저의 북마크 정보를 포함한 위키피디아 리스트 조회
      parameters:
      - name: user_id
        in: path
        description: 사용자 아이디
        required: true
        schema:
          type: string
          format: string
      - name: keyword
        in: query
        description: 검색 키워드
        required: false
        schema:
          type: string
          format: string
      responses:
        200:
          description: OK
          content:
            json:
              schema:
                type: object
                properties:
                  totalSize:
                    type: integer
                    description: 2d 전체 리스트 사이즈
                  size:
                    type: array
                    description: 2d 각각 원소 사이즈
                    items:
                      type: integer
                  wikis:
                    type: array
                    items:
                      type: array
                      items:                        
                        type: object
                        properties:
                          wikiId:
                            type: integer
                          title: 
                            type: string
                          createDate: 
                            type: string
                          creatorId: 
                            type: string
                          imgSrc: 
                            type: string
                          description: 
                            type: string
                          summary: 
                            type: string
                          source: 
                            type: string
                          bookmark: 
                            type: boolean
                          uks:
                            type: array
                            items:
                              type: string
  /users/{user_id}/wikis/bookmark:
    get:
      tags:
        - Wikipedia
      summary : 유저의 북마크 위키피디아 리스트 조회
      description : 유저가 북마크 선택한 위키피디아 리스트 조회
      parameters:
      - name: user_id
        in: path
        description: 사용자 아이디
        required: true
        schema:
          type: string
          format: string
      - name: keyword
        in: query
        description: 검색 키워드
        required: false
        schema:
          type: string
          format: string
      responses:
        200:
          description: OK
          content:
            json:
              schema:
                type: object
                properties:
                  totalSize:
                    type: integer
                    description: 2d 전체 리스트 사이즈
                  size:
                    type: array
                    description: 2d 각각 원소 사이즈
                    items:
                      type: integer
                  wikis:
                    type: array
                    items:
                      type: array
                      items:                        
                        type: object
                        properties:
                          wikiId:
                            type: integer
                          title: 
                            type: string
                          createDate: 
                            type: string
                          creatorId: 
                            type: string
                          imgSrc: 
                            type: string
                          description: 
                            type: string
                          summary: 
                            type: string
                          source: 
                            type: string
                          bookmark: 
                            type: boolean
                          uks:
                            type: array
                            items:
                              type: string
  /users/{user_id}/wikis/{wiki_id}/bookmark:
    put:
      tags:
        - Wikipedia
      summary : 유저의 위키피디아 북마크 추가
      description : 유저가 위키피디아에 북마크 추가
      parameters:
      - name: user_id
        in: path
        description: 사용자 아이디
        required: true
        schema:
          type: string
          format: string
      - name: wiki_id
        in: path
        description: 위키피디아 아이디
        required: true
        schema:
          type: string
          format: string
      responses:
        200:
          description: OK
          content:
            json:
              schema:
                type: object
                example:
                  isSuccess: true
    delete:
      tags:
        - Wikipedia
      summary : 유저의 위키피디아 북마크 삭제
      description : 유저가 위키피디아에 북마크 삭제
      parameters:
      - name: user_id
        in: path
        description: 사용자 아이디
        required: true
        schema:
          type: string
          format: string
      - name: wiki_id
        in: path
        description: 위키피디아 아이디
        required: true
        schema:
          type: string
          format: string
      responses:
        200:
          description: OK
          content:
            json:
              schema:
                type: object
                example:
                  isSuccess: true
  /users/{user_id}/wikis/{wiki_id}/hit:
    post:
      tags:
      - Wikipedia
      summary: 위키 조회수 업데이트
      description: 해당 위키의 조회수를 증가시키고 LRS 저장
      parameters:
      - name: user_id
        in: path
        description: 유저아이디
        required: true
        schema:
          type: string
          format: string
      - name: wiki_id
        in: path
        description: 위키아이디
        required: true
        schema:
          type: integer
          format: integer
      responses:
        200:
          description: OK
          content:
            json:
              schema:
                type: object
                example:
                  isSuccess: true                  
  /users/{user_id}/wikis/{wiki_id}/quit:
    post:
      tags:
      - Wikipedia
      summary: 위키 quit 업데이트
      description: 해당 위키 quit LRS 저장
      parameters:
      - name: user_id
        in: path
        description: 유저아이디
        required: true
        schema:
          type: string
          format: string
      - name: wiki_id
        in: path
        description: 백과사전아이디
        required: true
        schema:
          type: integer
          format: integer
      - name: duration
        in: query
        description: 컨텐츠 소비 시간
        required: true
        schema:
          type: integer
          format: integer
      responses:
        200:
          description: OK
          content:
            json:
              schema:
                type: object
                example:
                  isSuccess: true
...